fix: pr issues
