+++ src/globals/api/apiCall.ts
@@ -5,7 +5,9 @@ import { LogError } from '@/globals/api/LogError'
  * @param object - Любой объект
  */
 function convertObjectToURLSearchParams(object: object) {
-  const filtered = Object.entries(object).filter(([_, value]) => value !== undefined && value !== null)
+  const filtered = Object.entries(object).filter(([_, value]) => {
+    return value !== undefined && value !== null
+  })
   return new URLSearchParams(filtered)
 }
 
@@ -81,9 +83,9 @@ export async function apiCall<TRequest extends APIRequest | false = false, TResp
 
   if (request && method === 'POST') {
     fetchInit.body = JSON.stringify(request)
-    fetchInit.headers = Object.assign(fetchInit.headers || {}, {
+    fetchInit.headers = {
       'Content-Type': 'application/json',
-    })
+    }
   }
 
   if (request && method === 'GET') {
